Improved EnergyController to throttle input execution and to execute listeners in separate threads.
